# Conflicts:
#	app/src/main/java/com/example/duantotnghiep_md27/Adapter/Cart_Adapter.java
#	app/src/main/java/com/example/duantotnghiep_md27/Fragment/Cart_Fragment.java
#	app/src/main/java/com/example/duantotnghiep_md27/Fragment/Home_Fragment.java
#	app/src/main/java/com/example/duantotnghiep_md27/MainActivity.java
#	app/src/main/res/layout/activity_main.xml

+++ app/src/main/java/com/example/duantotnghiep_md27/Model/Product_home.java
@@ -12,7 +12,7 @@ public class Product_home{
     @SerializedName("product_name")
     private String product_name;
     @SerializedName("price")
-    private double price;
+    private int price;
     @SerializedName("image_url")
     private String image_url;
     @SerializedName("description")
@@ -26,7 +26,7 @@ public class Product_home{
     private Category category;
     private List<ProductSize> product_Sizes;
 
-    public Product_home(String product_id, String product_name, double price, String image_url, String description, String category_id, String creation_date, int quantity, Category category, List<ProductSize> product_Sizes) {
+    public Product_home(String product_id, String product_name, int price, String image_url, String description, String category_id, String creation_date, int quantity, Category category, List<ProductSize> product_Sizes) {
         this.product_id = product_id;
         this.product_name = product_name;
         this.price = price;
@@ -55,11 +55,11 @@ public class Product_home{
         this.product_name = product_name;
     }
 
-    public double getPrice() {
+    public int getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(int price) {
         this.price = price;
     }
